refactor: Indicator 컴포넌트 예시 파일에 추가해서 사용해보기
