+++ src/main.js
@@ -1,5 +1,7 @@
-import { 상품목록_레이아웃_로딩 } from "./components/productList/publish/상품목록_레이아웃_로딩.js";
-import { 상품목록_레이아웃_로딩완료 } from "./components/productList/publish/상품목록_레이아웃_로딩완료.js";
+// import { 상품목록_레이아웃_로딩 } from "./components/productList/publish/상품목록_레이아웃_로딩.js";
+// import { 상품목록_레이아웃_로딩완료 } from "./components/productList/publish/상품목록_레이아웃_로딩완료.js";
+import { 상품목록_레이아웃_로딩 } from "./components/productList/pages/상품목록_레이아웃_로딩.js";
+import { 상품목록_레이아웃_로딩완료 } from "./components/productList/pages/상품목록_레이아웃_로딩완료.js";
 import { 상품목록_레이아웃_카테고리_1Depth } from "./components/productList/publish/상품목록_레이아웃_카테고리_1Depth.js";
 import { 상품목록_레이아웃_카테고리_2Depth } from "./components/productList/publish/상품목록_레이아웃_카테고리_2Depth.js";
 import { 토스트 } from "./components/toast/토스트.js";
